Minor changes, some fixes and saving the state

+++ src/repos/setup.rs
@@ -1,17 +1,38 @@
-use crate::repos::database::SQLite;
+use std::path::Path;
 
 use super::{config::Config, errors::SetupError};
 use super::cache;
 
 pub fn setup(config: &mut Config) -> Result<(), SetupError> {
-    println!("Updating the database ...");
-    println!("Removing old {:?}", config.db);
-    std::fs::remove_file(&config.db)?;
-
-    config.sqlite = Some(SQLite::new(&config.db)?);
-    cache::dump_into_db(config)?;
+    println!("Syncing the database ...");
+    
+    if Path::new(&config.db).exists() {
+        println!("It seems you have an old database at {:?}", config.db);
+        println!("Do you want to override it? [y/N]");
+        
+        let mut input = String::new();
+        std::io::stdin()
+            .read_line(&mut input)
+            .expect("Could not read from stdin");
+        let input = input.trim().to_ascii_lowercase();
+        
+        match input.as_ref() {
+            "y" => {
+                println!("Removing old {:?}", config.db);
+                std::fs::remove_file(&config.db)?;
+                config.setup_db()?;
+                cache::populate_db(config)?;
+            }
+            _ => ()
+        }
+    } else {
+        config.setup()?;
+        config.setup_db()?;
+        cache::populate_db(config)?;
+    }
 
     let is_empty = config.cache.read_dir()?.next().is_none();
+
     if is_empty {
         println!("Consider do `opm update` before continue ...");
         return Ok(());
